Basic content of the box component working (middle)
Using negative margins and z-index means the velvet background will always look
like its lined up with the segment above or below it.

Currently if the box becomes bigger than the size of the velvet image, it will
have a seam where it repeats. This can probably be handled two ways:

1. Prevent the box from getting that big with a max-width, max-height
2. Use Gimp to make the velvet image seamless and repeatable
   (filters -> map -> make seamless)