Create the Pack components and tweak UI
This is basically the UI I had in mind.

I've followed the principle of progressive enhancement with regards to CSS grid.
Without any grid support, the elements should render in a standard column layout
(in some cases falling back to flexbox and mostly just standard display: block)

Then on desktops with CSS grid support, the UI is better spaced out.